make connection_store a PeerStore component
