+++ misc/peer-store/Cargo.toml
@@ -13,6 +13,7 @@ libp2p-core = { workspace = true }
 libp2p-swarm = { workspace = true }
 lru = "0.12.3"
 libp2p-identity = { workspace = true, optional = true }
+tracing = { workspace = true }
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
