+++ misc/peer-store/src/lib.rs
@@ -20,6 +20,7 @@
 mod behaviour;
 pub mod memory_store;
 mod store;
+pub mod connection_store;
 
 pub use behaviour::Behaviour;
 pub use store::Store;
